Thursday Morning Changes
So I fixed the problem with the player getting sick. Bennoni we were
close. It was GameObject.Find("player").GetComponent<Player>();.

We were getting the player object but not the script which contained the
sick variable itself.

Made some changes.
In the Player script:
*Renamed Sick to Ill.
*Made a sick() method which takes in an int parameter (the germ's id)
*Made a placeholder effect.
*Made it so that after 10 seconds of getting sick, the player's speed
increases, But not as much as before.

In the playerCollision
*Added a reference to Player script, named it 'me'
*fixed reference to enemy, named it 'germ'
*made a call to Player's new sick function and made it take germ.id as a
parameter.
*Made it so that Ill is true so if the player is sick he won't get
another illness on top of that. (Can change later if you guys want).